requirements.txt
Django>=4.2
python-telegram-bot>=20.0
asgiref>=3.7.0